+++ src/components/Chart/Chart.tsx
@@ -15,6 +15,7 @@ import {
 import { useTheme } from 'styled-components';
 
 import { ChartDatumBase, ChartProps } from './Chart.types';
+import { ChartTooltip } from './ChartTooltip';
 
 type HaloDotProps = { cx?: number; cy?: number };
 
@@ -118,7 +119,7 @@ export function Chart<TDatum extends ChartDatumBase = ChartDatumBase>(props: Cha
                     />
                 )}
 
-                <Tooltip cursor={{ fill: 'transparent' }} {...tooltipProps} />
+                <Tooltip cursor={{ fill: 'transparent' }} content={ChartTooltip} {...tooltipProps} />
                 {variant === 'bar+line' && (
                     <Legend align="center" iconSize={8} wrapperStyle={{ fontSize: 10 }} {...legendProps} />
                 )}
